pyproject.toml
[project]
name = "agentauditor"
version = "0.1.0"
description = "Runtime security agent that intercepts, evaluates, and blocks malicious agent actions before they execute"
requires-python = ">=3.10"
license = "Apache-2.0"
readme = "README.md"
keywords = ["ai", "security", "agents", "mcp", "guardrails", "owasp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "typer>=0.12",
    "rich>=13.0",
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]

[project.optional-dependencies]
mcp = ["fastmcp>=3.1.0"]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.50"]
llm = ["anthropic>=0.40", "openai>=1.50"]
all = ["fastmcp>=3.1.0", "anthropic>=0.40", "openai>=1.50"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.10",
]

[project.scripts]
agentauditor = "agentauditor.cli.app:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agentauditor"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
